Routing refactored again - tickets 770 and 768 completed
