+++ BCVaccineCard/BCVaccineCard/Helpers/Router/RouterWorker.swift
@@ -12,24 +12,16 @@ import UIKit
 enum CurrentPatientScenarios {
     case NoUsers
     case OneAuthUser
-    case OneUnauthUser
-    case MoreThanOneUnauthUser
-    case OneAuthUserAndOneUnauthUser
-    case OneAuthUserAndMoreThanOneUnauthUser
+//    case OneUnauthUser
+//    case MoreThanOneUnauthUser
+//    case OneAuthUserAndOneUnauthUser
+//    case OneAuthUserAndMoreThanOneUnauthUser
     
-    static func getCurrentScenario(authCount: Int, unauthCount: Int) -> Self {
-        if authCount == 0 && unauthCount == 0 {
+    static func getCurrentScenario(authCount: Int) -> Self {
+        if authCount == 0 {
             return .NoUsers
-        } else if authCount == 1 && unauthCount == 0 {
+        } else if authCount == 1 {
             return .OneAuthUser
-        } else if authCount == 0 && unauthCount == 1 {
-            return .OneUnauthUser
-        } else if authCount == 0 && unauthCount > 1 {
-            return .MoreThanOneUnauthUser
-        } else if authCount == 1 && unauthCount == 1 {
-            return .OneAuthUserAndOneUnauthUser
-        } else if authCount == 1 && unauthCount > 1 {
-            return .OneAuthUserAndMoreThanOneUnauthUser
         } else {
             return .NoUsers
         }
@@ -39,7 +31,7 @@ enum CurrentPatientScenarios {
 enum VaccineCardNumber {
     case NoCards
     case OneCard
-    case MultiplCards
+    case MultipleCards
     
     static func getCurrentNumber(cards: Int) -> Self {
         if cards == 0 {
@@ -47,7 +39,7 @@ enum VaccineCardNumber {
         } else if cards == 1 {
             return .OneCard
         } else {
-            return .MultiplCards
+            return .MultipleCards
         }
     }
 }
@@ -86,20 +78,20 @@ struct CurrentRecordsAndPassesStacks {
 enum RecordsFlowVCs {
     case HealthRecordsViewController
     case UsersListOfRecordsViewController(patient: Patient?)
-    case FetchHealthRecordsViewController
+//    case FetchHealthRecordsViewController
     case HealthRecordDetailViewController(patient: Patient?, dataSource: HealthRecordsDetailDataSource, userNumberHealthRecords: Int)
     case ProfileAndSettingsViewController
     case SecurityAndDataViewController
-    case GatewayFormViewController(rememberDetails: RememberedGatewayDetails, fetchType: GatewayFormViewControllerFetchType, gatewayInProgressDetails: GatewayInProgressDetails?)
+//    case GatewayFormViewController(rememberDetails: RememberedGatewayDetails, fetchType: GatewayFormViewControllerFetchType, gatewayInProgressDetails: GatewayInProgressDetails?)
     
     enum NonAssociatedVersion {
         case HealthRecordsViewController
         case UsersListOfRecordsViewController
-        case FetchHealthRecordsViewController
+//        case FetchHealthRecordsViewController
         case HealthRecordDetailViewController
         case ProfileAndSettingsViewController
         case SecurityAndDataViewController
-        case GatewayFormViewController
+//        case GatewayFormViewController
         
         func getIndexFromArray(array: [RecordsFlowVCs]) -> Int? {
             return array.map { $0.getNonAssociatedVersion }.firstIndex(of: self)
@@ -112,16 +104,16 @@ enum RecordsFlowVCs {
             return .HealthRecordsViewController
         case .UsersListOfRecordsViewController:
             return .UsersListOfRecordsViewController
-        case .FetchHealthRecordsViewController:
-            return .FetchHealthRecordsViewController
+//        case .FetchHealthRecordsViewController:
+//            return .FetchHealthRecordsViewController
         case .HealthRecordDetailViewController:
             return .HealthRecordDetailViewController
         case .ProfileAndSettingsViewController:
             return .ProfileAndSettingsViewController
         case .SecurityAndDataViewController:
             return .SecurityAndDataViewController
-        case .GatewayFormViewController:
-            return .GatewayFormViewController
+//        case .GatewayFormViewController:
+//            return .GatewayFormViewController
         }
     }
 }
@@ -196,20 +188,19 @@ class RouterWorker: NSObject {
         return StorageService.shared.fetchAuthenticatedPatient()
     }
     
-    private var getUnathenticatedPatients: [Patient]? {
-        return StorageService.shared.fetchUnauthenticatedPatients()
-    }
+//    private var getUnathenticatedPatients: [Patient]? {
+//        return StorageService.shared.fetchUnauthenticatedPatients()
+//    }
     
     private var currentPatientScenario: CurrentPatientScenarios {
-        let authPatientCount = StorageService.shared.fetchAuthenticatedPatient() != nil ? 1 : 0
-        let unauthPatientsCount = StorageService.shared.fetchUnauthenticatedPatients()?.count ?? 0
-        return CurrentPatientScenarios.getCurrentScenario(authCount: authPatientCount, unauthCount: unauthPatientsCount)
+        let authPatientCount = self.getAuthenticatedPatient != nil ? 1 : 0
+        return CurrentPatientScenarios.getCurrentScenario(authCount: authPatientCount)
     }
     
+    // Note: We really don't need this anymore, however, leaving it for when we potentially have to support dependants of an auth user
     private var userRecordsNavStyle: UsersListOfRecordsViewController.NavStyle {
-        let authPatientCount = StorageService.shared.fetchAuthenticatedPatient() != nil ? 1 : 0
-        let unauthPatientsCount = StorageService.shared.fetchUnauthenticatedPatients()?.count ?? 0
-        return authPatientCount + unauthPatientsCount > 1 ? .multiUser : .singleUser
+        let authPatientCount = self.getAuthenticatedPatient != nil ? 1 : 0
+        return authPatientCount > 1 ? .multiUser : .singleUser
     }
     
     private var currentNumberOfVaccineCards: VaccineCardNumber {
@@ -262,24 +253,14 @@ extension RouterWorker {
         guard values.affectedTabs.contains(.records) else { return [] }
         switch self.currentPatientScenario {
         case .NoUsers:
-            let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: false)
+            let vc = HealthRecordsViewController.constructHealthRecordsViewController()
             return [vc]
         case .OneAuthUser:
             guard let patient = self.getAuthenticatedPatient else {
-                let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: false)
+                let vc = HealthRecordsViewController.constructHealthRecordsViewController()
                 return [vc]
             }
             let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: true, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-            return [vc]
-        case .OneUnauthUser:
-            guard let patient = self.getUnathenticatedPatients?.first else {
-                let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: false)
-                return [vc]
-            }
-            let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: false, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-            return [vc]
-        case .MoreThanOneUnauthUser, .OneAuthUserAndOneUnauthUser, .OneAuthUserAndMoreThanOneUnauthUser:
-            let vc = HealthRecordsViewController.constructHealthRecordsViewController()
             return [vc]
         }
     }
@@ -293,7 +274,7 @@ extension RouterWorker {
     private func authFetchRecordsStack(values: ActionScenarioValues) -> [BaseViewController] {
         if values.currentTab == .records {
             switch self.currentPatientScenario {
-            case .NoUsers, .OneUnauthUser, .MoreThanOneUnauthUser:
+            case .NoUsers:
                 // Not possible here - after an authFetch, there has to be at least one Auth user, so do nothing
                 return []
             case .OneAuthUser:
@@ -302,13 +283,6 @@ extension RouterWorker {
                 guard let patient = values.recordFlowDetails?.actioningPatient else { return [] }
                 let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: true, navStyle: .singleUser, hasUpdatedUnauthPendingTest: true)
                 return [vc]
-            case .OneAuthUserAndOneUnauthUser, .OneAuthUserAndMoreThanOneUnauthUser:
-                // Stack should be HealthRecordsViewController, then UsersListOfRecordsViewController (after fetch is completed)
-                // Note - setting hasUpdatedUnauthPendingTest to false just in case unauth user has to check for background update for pending covid test
-                let vc1 = HealthRecordsViewController.constructHealthRecordsViewController()
-                guard let patient = values.recordFlowDetails?.actioningPatient else { return [vc1] }
-                let vc2 = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: true, navStyle: .multiUser, hasUpdatedUnauthPendingTest: false)
-                return [vc1, vc2]
             }
         }
         else {
@@ -323,7 +297,7 @@ extension RouterWorker {
             case .NoCards, .OneCard:
                 let vc = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: values.passesFlowDetails?.fedPassStringToOpen)
                 return [vc]
-            case .MultiplCards:
+            case .MultipleCards:
                 let vc1 = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: nil)
                 let vc2 = CovidVaccineCardsViewController.constructCovidVaccineCardsViewController(recentlyAddedCardId: values.passesFlowDetails?.recentlyAddedCardId, fedPassStringToOpen: values.passesFlowDetails?.fedPassStringToOpen)
                 return [vc1, vc2]
@@ -337,29 +311,13 @@ extension RouterWorker {
         if values.currentTab == .records {
             switch self.currentPatientScenario {
             case .NoUsers, .OneAuthUser:
-                // Not possible here - after a successful manual fetch, there has to be at least 1 unauth patient
+                // Not possible here - manual fetch is only for vaccine cards, won't show up in records tab
                 return []
-            case .OneUnauthUser:
-                // Stack should be UsersListOfRecordsViewController, then HealthRecordDetailViewController
-                guard let patient = values.recordFlowDetails?.actioningPatient else { return [] }
-                let vc1 = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: false, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-                let recordsCount = StorageService.shared.getHeathRecords().detailDataSource(patient: patient).count
-                guard let record = values.recordFlowDetails?.addedRecord else { return [vc1] }
-                let vc2 = HealthRecordDetailViewController.constructHealthRecordDetailViewController(dataSource: record, authenticatedRecord: false, userNumberHealthRecords: recordsCount, patient: patient)
-                return [vc1, vc2]
-            case .MoreThanOneUnauthUser, .OneAuthUserAndOneUnauthUser, .OneAuthUserAndMoreThanOneUnauthUser:
-                // Stack should be HealthRecordsViewController, UsersListOfRecordsViewController, then HealthRecordDetailViewController
-                let vc1 = HealthRecordsViewController.constructHealthRecordsViewController()
-                guard let patient = values.recordFlowDetails?.actioningPatient else { return [vc1] }
-                let authenticated = self.currentPatientScenario == .MoreThanOneUnauthUser ? false : true
-                let vc2 = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: authenticated, navStyle: .multiUser, hasUpdatedUnauthPendingTest: false)
-                guard let record = values.recordFlowDetails?.addedRecord else { return [vc1, vc2] }
-                let recordsCount = StorageService.shared.getHeathRecords().detailDataSource(patient: patient).count
-                let vc3 = HealthRecordDetailViewController.constructHealthRecordDetailViewController(dataSource: record, authenticatedRecord: false, userNumberHealthRecords: recordsCount, patient: patient)
-                return [vc1, vc2, vc3]
             }
-        } else {
-            return self.resetRecordsTab()
+        }
+        else {
+//            return self.resetRecordsTab()
+            return []
         }
     }
     
@@ -369,7 +327,7 @@ extension RouterWorker {
             case .NoCards, .OneCard:
                 let vc = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: values.passesFlowDetails?.fedPassStringToOpen)
                 return [vc]
-            case .MultiplCards:
+            case .MultipleCards:
                 if values.passesFlowDetails?.fedPassAddedFromHealthPassVC == true {
                     let vc1 = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: values.passesFlowDetails?.fedPassStringToOpen)
                     return [vc1]
@@ -388,29 +346,14 @@ extension RouterWorker {
         if values.currentTab == .records {
             guard values.affectedTabs.contains(.records) else { return [] }
             switch self.currentPatientScenario {
-            case .NoUsers:
-                // In this case, show initial fetch screen - stack should be FetchHealthRecordsViewController
-                let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: false)
-                return [vc]
-            case .OneAuthUser:
-                // This means that we have removed the only other unauth user and should show the auth user records by default UsersListOfRecordsViewController
-                // Note - hasUpdatedUnauthPendingTest doesnt matter here
-                guard let remainingAuthPatient = self.getAuthenticatedPatient else { return [] }
-                let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: remainingAuthPatient, authenticated: true, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-                return [vc]
-            case .OneUnauthUser:
-                // This means that we have removed an unauth user and should show the remaining unauth user records by default UsersListOfRecordsViewController
-                // Note - hasUpdatedUnauthPendingTest should be false here, to make sure that pending covid test result can be fetched in the background, just in case an update is required
-                guard let remainingUnauthPatient = self.getUnathenticatedPatients?.first else { return [] }
-                let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: remainingUnauthPatient, authenticated: false, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-                return [vc]
-            case .MoreThanOneUnauthUser, .OneAuthUserAndOneUnauthUser, .OneAuthUserAndMoreThanOneUnauthUser:
-                // In this case, just show the health records home screen with a list of folders
-                let vc = HealthRecordsViewController.constructHealthRecordsViewController()
-                return [vc]
+            case .NoUsers, .OneAuthUser:
+                // Not possible here - manual delete is only for vaccine cards, won't show up in records tab
+                return []
             }
-        } else {
-            return self.resetRecordsTab()
+        }
+        else {
+//            return self.resetRecordsTab()
+            return []
         }
     }
     
@@ -421,7 +364,7 @@ extension RouterWorker {
             case .NoCards, .OneCard:
                 let vc = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: nil)
                 return [vc]
-            case .MultiplCards:
+            case .MultipleCards:
                 let vc1 = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: nil)
                 let vc2 = CovidVaccineCardsViewController.constructCovidVaccineCardsViewController(recentlyAddedCardId: nil, fedPassStringToOpen: nil)
                 return [vc1, vc2]
@@ -435,24 +378,13 @@ extension RouterWorker {
         if values.currentTab == .records {
             switch self.currentPatientScenario {
             case .NoUsers:
-                let vc1 = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: false)
-                guard values.currentTab == .records else { return [vc1] }
-                let vc2 = ProfileAndSettingsViewController.constructProfileAndSettingsViewController()
-                return [vc1, vc2]
-            case .OneAuthUser, .OneAuthUserAndOneUnauthUser, .OneAuthUserAndMoreThanOneUnauthUser:
-                // This isn't possible - after logout, there is no auth user
-                return []
-            case .OneUnauthUser:
-                guard let remainingUnauthPatient = self.getUnathenticatedPatients?.first else { return [] }
-                let vc1 = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: remainingUnauthPatient, authenticated: false, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-                guard values.currentTab == .records else { return [vc1] }
-                let vc2 = ProfileAndSettingsViewController.constructProfileAndSettingsViewController()
-                return [vc1, vc2]
-            case .MoreThanOneUnauthUser:
                 let vc1 = HealthRecordsViewController.constructHealthRecordsViewController()
                 guard values.currentTab == .records else { return [vc1] }
                 let vc2 = ProfileAndSettingsViewController.constructProfileAndSettingsViewController()
                 return [vc1, vc2]
+            case .OneAuthUser:
+                // This isn't possible - after logout, there is no auth user
+                return []
             }
         } else {
             return self.resetRecordsTab()
@@ -467,7 +399,7 @@ extension RouterWorker {
                 guard values.currentTab == .healthPass else { return [vc1] }
                 let vc2 = ProfileAndSettingsViewController.constructProfileAndSettingsViewController()
                 return [vc1, vc2]
-            case .MultiplCards:
+            case .MultipleCards:
                 let vc1 = HealthPassViewController.constructHealthPassViewController(fedPassStringToOpen: nil)
                 let vc2 = CovidVaccineCardsViewController.constructCovidVaccineCardsViewController(recentlyAddedCardId: nil, fedPassStringToOpen: nil)
                 guard values.currentTab == .healthPass else { return [vc1, vc2] }
@@ -509,21 +441,13 @@ extension RouterWorker {
     private func resetRecordsTab() -> [BaseViewController] {
         switch self.currentPatientScenario {
         case .NoUsers:
-            let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: false)
+            let vc = HealthRecordsViewController.constructHealthRecordsViewController()
             return [vc]
         case .OneAuthUser:
             guard let patient = self.getAuthenticatedPatient else { return [] }
             let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: true, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
             return [vc]
-        case .OneUnauthUser:
-            guard let patient = self.getUnathenticatedPatients?.first else { return [] }
-            let vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: false, navStyle: .singleUser, hasUpdatedUnauthPendingTest: false)
-            return [vc]
-        case .MoreThanOneUnauthUser, .OneAuthUserAndOneUnauthUser, .OneAuthUserAndMoreThanOneUnauthUser:
-            let vc = HealthRecordsViewController.constructHealthRecordsViewController()
-            return [vc]
         }
-        
     }
     
     private func resetPassesTab() -> [BaseViewController] {
@@ -555,17 +479,17 @@ extension RouterWorker {
             case .UsersListOfRecordsViewController(patient: let patient):
                 guard let patient = patient else { return }
                 vc = UsersListOfRecordsViewController.constructUsersListOfRecordsViewController(patient: patient, authenticated: patient.authenticated, navStyle: self.userRecordsNavStyle, hasUpdatedUnauthPendingTest: false)
-            case .FetchHealthRecordsViewController:
-                let hasRecords = !(StorageService.shared.getHeathRecords().isEmpty)
-                vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: hasRecords)
+//            case .FetchHealthRecordsViewController:
+//                let hasRecords = !(StorageService.shared.getHeathRecords().isEmpty)
+//                vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: hasRecords)
             case .HealthRecordDetailViewController(patient: let patient, dataSource: let dataSource, userNumberHealthRecords: let userNumberHealthRecords):
                 vc = HealthRecordDetailViewController.constructHealthRecordDetailViewController(dataSource: dataSource, authenticatedRecord: dataSource.isAuthenticated, userNumberHealthRecords: userNumberHealthRecords, patient: patient)
             case .ProfileAndSettingsViewController:
                 vc = ProfileAndSettingsViewController.constructProfileAndSettingsViewController()
             case .SecurityAndDataViewController:
                 vc = SecurityAndDataViewController.constructSecurityAndDataViewController()
-            case .GatewayFormViewController(rememberDetails: let rememberDetails, fetchType: let fetchType, gatewayInProgressDetails: let currentProgress):
-                vc = GatewayFormViewController.constructGatewayFormViewController(rememberDetails: rememberDetails, fetchType: fetchType, currentProgress: currentProgress)
+//            case .GatewayFormViewController(rememberDetails: let rememberDetails, fetchType: let fetchType, gatewayInProgressDetails: let currentProgress):
+//                vc = GatewayFormViewController.constructGatewayFormViewController(rememberDetails: rememberDetails, fetchType: fetchType, currentProgress: currentProgress)
             }
             if let vc = vc {
                 newVCStack.append(vc)
@@ -580,21 +504,18 @@ extension RouterWorker {
         case .NoUsers:
             // Check if currentRecordsStack contains healthRecordsVC or userListOfRecordsVC or recordsDetailVC - if so, then remove those from the stack. Make sure fetchVC is the base VC here
             var recordsStack = currentRecordsStack
-            if let index = RecordsFlowVCs.NonAssociatedVersion.HealthRecordsViewController.getIndexFromArray(array: recordsStack) {
-                recordsStack.remove(at: index)
-            }
             if let index = RecordsFlowVCs.NonAssociatedVersion.UsersListOfRecordsViewController.getIndexFromArray(array: recordsStack) {
                 recordsStack.remove(at: index)
             }
             if let index = RecordsFlowVCs.NonAssociatedVersion.HealthRecordDetailViewController.getIndexFromArray(array: recordsStack) {
                 recordsStack.remove(at: index)
             }
-            if let index = RecordsFlowVCs.NonAssociatedVersion.FetchHealthRecordsViewController.getIndexFromArray(array: recordsStack) {
-                let fetchVC = recordsStack.remove(at: index)
-                recordsStack.insert(fetchVC, at: 0)
+            if let index = RecordsFlowVCs.NonAssociatedVersion.HealthRecordsViewController.getIndexFromArray(array: recordsStack) {
+                let healthRecordsVC = recordsStack.remove(at: index)
+                recordsStack.insert(healthRecordsVC, at: 0)
             } else {
-                let fetchVC = RecordsFlowVCs.FetchHealthRecordsViewController
-                recordsStack.insert(fetchVC, at: 0)
+                let healthRecordsVC = RecordsFlowVCs.HealthRecordsViewController
+                recordsStack.insert(healthRecordsVC, at: 0)
             }
             return recordsStack
 //        case .OneAuthUser:
