+++ BCVaccineCard/BCVaccineCard/Screens/HealthRecordsFlow/FetchHealthRecordsViewController/FetchHealthRecordsViewController.swift
@@ -5,6 +5,8 @@
 //  Created by Connor Ogilvie on 2021-11-29.
 // TODO: This will contain a table view with the two cells for immunization record and test results. This view controller will be shown automatically on top of the healthRecordsViewController when no health records exist - will have to show a loading indicator
 
+// TODO: CONNOR: Delete this view controller
+
 import UIKit
 
 class FetchHealthRecordsViewController: BaseViewController {
@@ -36,9 +38,9 @@ class FetchHealthRecordsViewController: BaseViewController {
     private var dataSource: [DataSource] = [.recordType(type: .covidImmunizationRecord),
                                             .recordType(type: .covidTestResult)]
     
-    override var getRecordFlowType: RecordsFlowVCs? {
-        return .FetchHealthRecordsViewController
-    }
+//    override var getRecordFlowType: RecordsFlowVCs? {
+//        return .FetchHealthRecordsViewController
+//    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
