+++ BCVaccineCard/BCVaccineCard/Screens/HealthRecordsFlow/UsersListOfRecordsViewController/UsersListOfRecordsViewController.swift
@@ -165,10 +165,10 @@ extension UsersListOfRecordsViewController {
         
         if style == .singleUser {
             self.navigationItem.setHidesBackButton(true, animated: false)
-            let addButton = NavButton(title: nil,
-                      image: UIImage(named: "add-circle-btn"), action: #selector(self.showAddRecord),
-                                      accessibility: Accessibility(traits: .button, label: "", hint: "")) // TODO:
-            buttons.append(addButton)
+//            let addButton = NavButton(title: nil,
+//                      image: UIImage(named: "add-circle-btn"), action: #selector(self.showAddRecord),
+//                                      accessibility: Accessibility(traits: .button, label: "", hint: "")) // TODO:
+//            buttons.append(addButton)
             let settingsButton = NavButton(title: nil,
                       image: UIImage(named: "nav-settings"), action: #selector(self.showSettings),
                                            accessibility: Accessibility(traits: .button, label: "", hint: "")) // TODO:
@@ -193,10 +193,11 @@ extension UsersListOfRecordsViewController {
                                                backButtonHintString: nil)
     }
     // This screen has to have health records by rule (with the exception being a screen state issue, but that is a separate bug)
-    @objc func showAddRecord() {
-        let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: true)
-        self.navigationController?.pushViewController(vc, animated: true)
-    }
+    // TODO: CONNOR: Remove this
+//    @objc func showAddRecord() {
+//        let vc = FetchHealthRecordsViewController.constructFetchHealthRecordsViewController(hasHealthRecords: true)
+//        self.navigationController?.pushViewController(vc, animated: true)
+//    }
     
     @objc func showSettings() {
         UIImpactFeedbackGenerator(style: .light).impactOccurred()
@@ -296,7 +297,8 @@ extension UsersListOfRecordsViewController {
             // Don't check for test result to update again here, as this was causing an infinite loop
             // TODO: We should really refactor the way screens are being updated due to storage updates, as it will cause issues in the future with edge cases, causing us to create numerous hot-fixes such as this, resulting in messy and hard to maintain code
         } else {
-            self.checkForTestResultsToUpdate(ds: self.dataSource)
+            // NOTE: We don't need to do this anymore - should check with the team though
+//            self.checkForTestResultsToUpdate(ds: self.dataSource)
         }
     }
     
@@ -407,31 +409,32 @@ extension UsersListOfRecordsViewController {
         NotificationCenter.default.post(name: .protectedWordRequired, object: nil, userInfo: userInfo)
     }
     
-    private func checkForTestResultsToUpdate(ds: [HealthRecordsDetailDataSource]) {
-        for (indexPathRow, record) in ds.enumerated() {
-            switch record.type {
-            case .covidTestResultRecord(model: let model):
-                let listOfStatuses = record.records.map { ($0.status, $0.date) }
-                for (index, data) in listOfStatuses.enumerated() {
-                    if data.0 == CovidTestResult.pending.rawValue {
-                        
-                        guard
-                            let patient = model.patient,
-                            let dateOfBirth = patient.birthday?.yearMonthDayString,
-                            let phn = patient.phn,
-                            let collectionDatePresentableFormat = listOfStatuses[index].1,
-                            let collectionDate = Date.Formatter.monthDayYearDate.date(from: collectionDatePresentableFormat)?.yearMonthDayString, model.authenticated == false
-                        else { return }
-                        
-                        let model = GatewayTestResultRequest(phn: phn, dateOfBirth: dateOfBirth, collectionDate: collectionDate)
-                        tableView.cellForRow(at: IndexPath(row: indexPathRow, section: 0))?.startLoadingIndicator(backgroundColor: .clear, containerSize: 20, size: 8)
-                        backgroundWorker?.getTestResult(model: model, executingVC: self, row: indexPathRow)
-                    }
-                }
-            default: print("")
-            }
-        }
-    }
+    // NOTE: We won't need this anymore I don't believe - double check
+//    private func checkForTestResultsToUpdate(ds: [HealthRecordsDetailDataSource]) {
+//        for (indexPathRow, record) in ds.enumerated() {
+//            switch record.type {
+//            case .covidTestResultRecord(model: let model):
+//                let listOfStatuses = record.records.map { ($0.status, $0.date) }
+//                for (index, data) in listOfStatuses.enumerated() {
+//                    if data.0 == CovidTestResult.pending.rawValue {
+//
+//                        guard
+//                            let patient = model.patient,
+//                            let dateOfBirth = patient.birthday?.yearMonthDayString,
+//                            let phn = patient.phn,
+//                            let collectionDatePresentableFormat = listOfStatuses[index].1,
+//                            let collectionDate = Date.Formatter.monthDayYearDate.date(from: collectionDatePresentableFormat)?.yearMonthDayString, model.authenticated == false
+//                        else { return }
+//
+//                        let model = GatewayTestResultRequest(phn: phn, dateOfBirth: dateOfBirth, collectionDate: collectionDate)
+//                        tableView.cellForRow(at: IndexPath(row: indexPathRow, section: 0))?.startLoadingIndicator(backgroundColor: .clear, containerSize: 20, size: 8)
+//                        backgroundWorker?.getTestResult(model: model, executingVC: self, row: indexPathRow)
+//                    }
+//                }
+//            default: print("")
+//            }
+//        }
+//    }
     
     private func performBCSCLogin() {
         self.showLogin(initialView: .Auth, sourceVC: .UserListOfRecordsVC) { [weak self] authenticated in
